requirements.txt
APScheduler==3.6.3
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
decorator==4.4.2
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
oauth2client==4.0.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-telegram-bot==13.1
pytz==2020.4
requests==2.25.0
rsa==4.6
six==1.15.0
tornado==6.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.2
